FSA-002 - IA local com Ollama

+++ src/main/kotlin/br/com/finsavior/monolith/finsavior_monolith/model/entity/ChatMessageHistory.kt
@@ -21,7 +21,7 @@ data class ChatMessageHistory(
     val chatMessageId: Long,
 
     @Column(name = "tokens_used")
-    val tokensUsed: Long,
+    val tokensUsed: Int,
 
     @Column(name = "created_at")
     val createdAt: LocalDateTime = LocalDateTime.now()
